requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
google-generativeai
